src/main/java/com/inventine/model/Payout.java
package com.inventine.model;

import java.sql.Timestamp;

public class Payout {

    private long amount;
    private int transactionId;
    private Timestamp createdAt;

    public long getAmount() {
        return amount;
    }

    public void setAmount(long amount) {
        this.amount = amount;
    }

    public int getTransactionId() {
        return transactionId;
    }

    public void setTransactionId(int transactionId) {
        this.transactionId = transactionId;
    }

    public Timestamp getCreatedAt() {
        return createdAt;
    }

    public void setCreatedAt(Timestamp createdAt) {
        this.createdAt = createdAt;
    }

}
